+++ crowdfunding/users/views.py
@@ -4,8 +4,10 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from .models import CustomUser
-from .serializers import CustomUserSerializer, CustomUserDetailSerializer
-
+from .serializers import CustomUserSerializer
+###########
+# Creation of a user profile. No authentication required.
+###########
 class CustomUserList(APIView):
     def get(self, request):
         users = CustomUser.objects.all()
@@ -19,7 +21,10 @@ class CustomUserList(APIView):
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors)
-    
+
+##########
+# This class and functions allows a user to edit their profiles details
+##########  
 
 class CustomUserDetail(APIView):
     def get_object(self, pk):
@@ -35,28 +40,48 @@ class CustomUserDetail(APIView):
 
     def put(self, request, pk):
         user = self.get_object(pk)
-        data = request.data
-        serializer = CustomUserDetailSerializer(
+        if user ==  request.user:
+            data = request.data
+            serializer = CustomUserSerializer(
             instance=user,
             data=data,
             partial=True
         )
-        if serializer.is_valid():
-            serializer.save()
+            if serializer.is_valid():
+                serializer.save()
+                return Response(
+                    serializer.data,
+                    status=status.HTTP_200_OK
+                )
             return Response(
-                serializer.data,
-                status=status.HTTP_200_OK
+                serializer.errors,
+                status=status.HTTP_400_BAD_REQUEST
             )
         return Response(
-            serializer.errors,
-            status=status.HTTP_400_BAD_REQUEST
+            status=status.HTTP_401_UNAUTHORIZED
         )
+    
 
     # DELETE request
     def delete(self, request, pk):
         user = self.get_object(pk)
-        user.delete()
+        if user == request.user:
+            user.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT
+            )
         return Response(
-            status=status.HTTP_204_NO_CONTENT
+            status=status.HTTP_401_UNAUTHORIZED
         )
 
+class CustomUserDetailByUsername(APIView):
+    
+    def get_object(self, username):
+        try:
+            return CustomUser.objects.get(username=username)
+        except CustomUser.DoesNotExist:
+            raise Http404
+    
+    def get(self, request, username):
+        user = self.get_object(username)
+        serializer = CustomUserSerializer(user)
+        return Response(serializer.data)
